requirements.txt
pyyaml
vosk
wave
ffmpeg-python
bs4
botocore
boto3
numpy==1.26.4
whisperx
python-dotenv
git+https://github.com/pyannote/pyannote-audio.git@3.1.1
mutagen
pydub
seaborn
fastapi
openai
pydantic
uvicorn
pysrt
pandas
faiss-cpu
tqdm